created DataView folder and moved all related components to there
