+++ sites/geohub/src/components/DataView.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
   import { page } from '$app/stores'
   import type { DataCategory, DataOrderType, DataSortingColumn, StacItemFeatureCollection } from '$lib/types'
-  import DataCard from './DataCard.svelte'
+  import DataCard from '$components/DataView/DataCard.svelte'
   import { map, indicatorProgress } from '$stores'
-  import TextFilter from './controls/TextFilter.svelte'
-  import Notification from './controls/Notification.svelte'
+  import TextFilter from '$components/DataView/TextFilter.svelte'
+  import Notification from '$components/controls/Notification.svelte'
   import { SEARCH_PAGINATION_LIMIT, DataCategories, DatasetSearchQueryParams } from '$lib/constants'
-  import DataCategoryCardList from './DataCategoryCardList.svelte'
-  import Breadcrumbs from './controls/Breadcrumbs.svelte'
+  import DataCategoryCardList from '$components/DataView/DataCategoryCardList.svelte'
+  import Breadcrumbs from '$components/controls/Breadcrumbs.svelte'
   import type { Tag } from '$lib/types/Tag'
 
   export let headerHeight: number = undefined
