+++ sites/geohub/src/components/DataView/DataCategoryCard.svelte
