+++ sites/geohub/src/components/DataView/MiniMap.svelte
@@ -118,8 +118,8 @@
 </div>
 
 <style lang="scss">
-  @use '../styles/undp-design/base-minimal.min.css';
-  @use '../styles/undp-design/loader.min.css';
+  @use '../../styles/undp-design/base-minimal.min.css';
+  @use '../../styles/undp-design/loader.min.css';
 
   .map-container {
     position: relative;
